Moved classes back to where they were
