+++ spring-cloud-dataflow-server-core/src/main/java/org/springframework/cloud/dataflow/server/service/impl/ResourceMixin.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.springframework.cloud.dataflow.core;
+package org.springframework.cloud.dataflow.server.service.impl;
 
 import com.fasterxml.jackson.annotation.JsonValue;
